Refactor create_person
